# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: InsightForge application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [Storage]
      summary: Request a private object upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadUrlRequest" }
      responses:
        "200":
          description: Signed upload URL.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadUrlResponse" }
        "400":
          description: Invalid upload metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          description: Authentication required.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [Storage]
      summary: Serve a private uploaded object to its owner
      parameters:
        - { name: objectPath, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Object content. }
        "401": { description: Authentication required. }
        "403": { description: Not the object owner. }
        "404": { description: File not found. }
...
  /reports/generate:
    post:
      operationId: generateReport
      tags: [Reports]
      summary: Generate and save a report from computed dataset results
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportRequest" }
      responses:
        "201":
          description: Saved report.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "404": { description: Dataset not found. }
  /reports/{reportId}:
    get:
      operationId: getReport
      tags: [Reports]
      summary: Get a saved report
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Report details.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
        "404": { description: Report not found. }
  /reports/{reportId}/download:
    get:
      operationId: downloadReport
      tags: [Reports]
      summary: Download a saved report as HTML
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Self-contained HTML report.
          content:
            text/html:
              schema: { type: string }
        "404": { description: Report not found. }
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        size: { type: integer, minimum: 1, maximum: 104857600 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata: { $ref: "#/components/schemas/UploadUrlRequest" }
...
    ReportRequest:
      type: object
      required: [datasetId]
      properties:
        datasetId: { type: string, format: uuid }
        experimentId: { type: string, format: uuid }
    Report:
      type: object
      required: [id, datasetId, title, content, createdAt]
      properties:
        id: { type: string, format: uuid }
        datasetId: { type: string, format: uuid }
        title: { type: string }
        content: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
...